LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.meta/script-src-self/worklet-audio.https-expected.txt

CONSOLE MESSAGE: Refused to load https://www1.web-platform.test:9443/common/security-features/subresource/worker.py?redirection=keep-origin&action=purge&key=d76f468a-9132-40b2-9c18-affff062a0fe&path=%2Fmixed-content because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://www1.web-platform.test:9443/common/security-features/subresource/worker.py?redirection=no-redirect&action=purge&key=5cf78223-c7a5-4f3a-a131-c80bec8e21ec&path=%2Fmixed-content because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://www1.web-platform.test:9443/common/security-features/subresource/worker.py?redirection=swap-origin&action=purge&key=38871af4-777b-4518-b331-10e011532044&path=%2Fmixed-content because it does not appear in the script-src directive of the Content Security Policy.

PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and keep-origin redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and keep-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and no-redirect redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and no-redirect redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and keep-origin redirection from https context.
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and keep-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and no-redirect redirection from https context.
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and no-redirect redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and swap-origin redirection from https context.
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and swap-origin redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects blocked for worklet-audio to same-https origin and swap-origin redirection from https context. assert_unreached: main promise resolved unexpectedly Reached unreachable code
FAIL Content Security Policy: Expects blocked for worklet-audio to same-https origin and swap-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
